requirements.txt
jsonschema
pydantic
pyyaml